added deploy scripts for dovecot and haraka
- check for cert file before deploy

+++ provision-letsencrypt.sh
@@ -15,14 +15,14 @@ install_letsencrypt()
 	local _installer="$STAGE_MNT/acme.sh"
 	fetch -o $_installer https://raw.githubusercontent.com/Neilpang/acme.sh/master/acme.sh
 	sed -i.bak -e '/^DEFAULT_INSTALL_HOME=/ s/=.*$/="\/data"/' $_installer
-	stage_exec sh /acme.sh --install --home=/data
+	stage_exec sh /acme.sh --install
 }
 
 # shellcheck disable=SC2120
 install_deploy_haproxy()
 {
 	# shellcheck disable=SC2154
-	tee "$_deploy/haproxy" <<EO_LE_HAPROXY
+	tee "$_deploy/haproxy" <<'EO_LE_HAPROXY'
 #!/bin/sh
 
 #returns 0 means success, otherwise error.
@@ -34,6 +34,11 @@ haproxy_deploy() {
   _ccert="$3"
   _cca="$4"
   _cfullchain="$5"
+
+  if [ ! -f $_ccert ]; then
+    _err "missing certificate"
+    exit 2
+  fi
 
   local _tmp="/tmp/${_cdomain}.pem"
   cat $_ckey $_ccert $_cfullchain > $_tmp
@@ -63,9 +68,110 @@ haproxy_deploy() {
 EO_LE_HAPROXY
 }
 
+# shellcheck disable=SC2120
+install_deploy_dovecot()
+{
+	# shellcheck disable=SC2154
+	tee "$_deploy/dovecot" <<'EO_LE_DOVECOT'
+#!/bin/sh
+
+#domain keyfile certfile cafile fullchain
+dovecot_deploy() {
+  _cdomain="$1"
+  _ckey="$2"
+  _ccert="$3"
+  _cca="$4"
+  _cfullchain="$5"
+
+  if [ ! -f $_ccert ]; then
+    _err "missing certificate"
+    exit 2
+  fi
+
+  local _tmp_crt="/tmp/dovecot-cert-${_cdomain}.pem"
+  cat $_ccert $_cfullchain > $_tmp_crt
+  if [ -s "$_tmp_crt" ]; then
+    _debug "$_tmp_crt created"
+    local _installed="/data/dovecot/etc/ssl/certs/dovecot.pem"
+    if diff -q $_tmp_crt $_installed; then
+      _debug "cert is the same, skip deploy"
+      return 0
+    fi
+
+    cp $_tmp $_installed || return 1
+    cp $_ckey /data/dovecot/etc/ssl/private/dovecot.pem || return 1
+    if [ -s "$_installed" ]; then
+      rm $_tmp
+      _debug "restarting dovecot"
+      jexec dovecot service dovecot restart
+      return 0
+    fi
+
+    _err "install to $_installed failed"
+    return 1
+  fi
+
+  _err "Unable to create $_tmp"
+  return 1
+}
+EO_LE_DOVECOT
+}
+
+# shellcheck disable=SC2120
+install_deploy_haraka()
+{
+	# shellcheck disable=SC2154
+	tee "$_deploy/haraka" <<'EO_LE_HARAKA'
+#!/bin/sh
+
+#returns 0 means success, otherwise error.
+
+#domain keyfile certfile cafile fullchain
+haraka_deploy() {
+  _cdomain="$1"
+  _ckey="$2"
+  _ccert="$3"
+  _cca="$4"
+  _cfullchain="$5"
+
+  if [ ! -f $_ccert ]; then
+    _err "missing certificate"
+    exit 2
+  fi
+
+  local _tmp="/tmp/${_cdomain}.pem"
+  cat $_ckey $_ccert $_cfullchain > $_tmp
+  if [ -s "$_tmp" ]; then
+    _debug "$_tmp created"
+    local _installed="/data/haraka/config/tls_cert.pem"
+    if diff -q $_tmp $_installed; then
+      _debug "cert is the same, skip deploy"
+      return 0
+    fi
+
+    cp $_tmp $_installed || return 1
+    cp $_ckey /data/haraka/config/tls_key.pem || return 1
+
+    if [ -s "$_installed" ]; then
+      rm $_tmp
+      _debug "restarting haraka"
+      service jail restart haraka
+      return 0
+    fi
+
+    _err "install to $_installed failed"
+    return 1
+  fi
+
+  _err "Unable to create $_tmp"
+  return 1
+}
+EO_LE_HARAKA
+}
+
 install_deploy_scripts()
 {
-	local _deploy="$ZFS_DATA_MNT/letsencrypt/deploy"
+	export _deploy="$ZFS_DATA_MNT/letsencrypt/deploy"
 	if [ ! -d $_deploy ]; then
 		mkdir $_deploy || exit
 	fi
